project_2/requirements.txt
numpy
scipy
Pillow 
scikit-image 
opencv-python==3.4.2.16
matplotlib
opencv-contrib-python==3.4.2.16
